Unify the `Variable` types
