+++ crates/espanso-config/src/counter.rs
@@ -19,12 +19,12 @@
 
 use std::sync::atomic::{AtomicI32, Ordering};
 
+use shinran_types::StructId;
+
 thread_local! {
   // TODO: if thread local, we probably don't need an atomic
   static STRUCT_COUNTER: AtomicI32 = const { AtomicI32::new(0) };
 }
-
-pub type StructId = i32;
 
 /// Some structs need a unique id.
 /// In order to generate it, we use an atomic static variable
