+++ crates/espanso-config/src/matches/group/loader/yaml/mod.rs
@@ -24,14 +24,15 @@ use crate::{
     error::{ErrorRecord, NonFatalErrorSet},
     matches::{
         group::{path::resolve_imports, LoadedMatchFile},
-        ImageEffect, Match, Params, RegexCause, TextFormat, TextInjectMode, UpperCasingStyle,
-        Value, Variable, WordBoundary,
+        ImageEffect, Match, RegexCause, TextFormat, TextInjectMode, UpperCasingStyle, Variable,
+        WordBoundary,
     },
 };
 use anyhow::{anyhow, bail, Context, Result};
 use lazy_static::lazy_static;
 use parse::YAMLMatchFile;
 use regex::{Captures, Regex};
+use shinran_types::{Params, Value, VarType};
 
 use self::{
     parse::{YAMLMatch, YAMLVariable},
@@ -279,7 +280,7 @@ pub fn try_convert_into_match(
         let vars = vec![Variable {
             id: next_id(),
             name: "form1".to_owned(),
-            var_type: "form".to_owned(),
+            var_type: VarType::Form,
             params,
             ..Default::default()
         }];
@@ -320,10 +321,23 @@ pub fn try_convert_into_variable(
     yaml_var: YAMLVariable,
     use_compatibility_mode: bool,
 ) -> Result<(Variable, Vec<Warning>)> {
+    let var_type = match &yaml_var.var_type[..] {
+        "date" => VarType::Date,
+        "dummy" => VarType::Echo,
+        "echo" => VarType::Echo,
+        "form" => VarType::Form,
+        "match" => VarType::Match,
+        "random" => VarType::Random,
+        "script" => VarType::Script,
+        "shell" => VarType::Shell,
+        "mock" => VarType::Mock,
+        // "global" => VarType::Global,
+        _ => return Err(anyhow!("unknown variable type: {:?}", yaml_var.var_type)),
+    };
     Ok((
         Variable {
             name: yaml_var.name,
-            var_type: yaml_var.var_type,
+            var_type,
             params: convert_params(yaml_var.params)?,
             id: next_id(),
             inject_vars: !use_compatibility_mode && yaml_var.inject_vars.unwrap_or(true),
@@ -336,10 +350,7 @@ pub fn try_convert_into_variable(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{
-        matches::{Match, Params, Value},
-        util::tests::use_test_directory,
-    };
+    use crate::{matches::Match, util::tests::use_test_directory};
     use std::{ffi::OsString, fs::create_dir_all};
 
     fn create_match_with_warnings(
@@ -632,7 +643,7 @@ mod tests {
                     vars: vec![Variable {
                         id: 0,
                         name: "form1".to_string(),
-                        var_type: "form".to_string(),
+                        var_type: VarType::Form,
                         params,
                         ..Default::default()
                     }],
@@ -669,7 +680,7 @@ mod tests {
                     vars: vec![Variable {
                         id: 0,
                         name: "form1".to_string(),
-                        var_type: "form".to_string(),
+                        var_type: VarType::Form,
                         params,
                         ..Default::default()
                     }],
@@ -708,7 +719,7 @@ mod tests {
                     vars: vec![Variable {
                         id: 0,
                         name: "form1".to_string(),
-                        var_type: "form".to_string(),
+                        var_type: VarType::Form,
                         params,
                         ..Default::default()
                     }],
@@ -725,7 +736,6 @@ mod tests {
         params.insert("param1".to_string(), Value::Bool(true));
         let vars = vec![Variable {
             name: "var1".to_string(),
-            var_type: "test".to_string(),
             params,
             ..Default::default()
         }];
@@ -736,7 +746,7 @@ mod tests {
         replace: "world"
         vars:
           - name: var1
-            type: test
+            type: mock
             params:
               param1: true
         "#
@@ -762,13 +772,11 @@ mod tests {
         let vars = vec![
             Variable {
                 name: "var1".to_string(),
-                var_type: "test".to_string(),
                 depends_on: vec!["test".to_owned()],
                 ..Default::default()
             },
             Variable {
                 name: "var2".to_string(),
-                var_type: "test".to_string(),
                 inject_vars: false,
                 ..Default::default()
             },
@@ -780,10 +788,10 @@ mod tests {
         replace: "world"
         vars:
           - name: var1
-            type: test
+            type: mock
             depends_on: ["test"]
           - name: var2
-            type: "test"
+            type: "mock"
             inject_vars: false
         "#
             )
@@ -807,7 +815,6 @@ mod tests {
     fn vars_no_params_maps_correctly() {
         let vars = vec![Variable {
             name: "var1".to_string(),
-            var_type: "test".to_string(),
             params: Params::new(),
             ..Default::default()
         }];
@@ -818,7 +825,7 @@ mod tests {
         replace: "world"
         vars:
           - name: var1
-            type: test
+            type: mock
         "#
             )
             .unwrap(),
@@ -862,7 +869,7 @@ mod tests {
 
       global_vars:
         - name: "var1"
-          type: "test"
+          type: "mock"
 
       matches:
         - trigger: "hello"
@@ -885,7 +892,6 @@ mod tests {
 
             let vars = vec![Variable {
                 name: "var1".to_string(),
-                var_type: "test".to_string(),
                 params: Params::new(),
                 ..Default::default()
             }];
