+++ crates/shinran-types/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "shinran-types"
+version = "0.1.0"
+edition.workspace = true
+
+[dependencies]
+enum-as-inner.workspace = true
+
+[lints]
+workspace = true
